// src/infraestructure/interfaces/interfacePrueba.tsx
// Generated by https://quicktype.io

export interface MedicinaR {
    rxtermsProperties: {
        displayName: string,
        synonym: string,
        fullName: string, 
        fullGenericName: string,
        strength: string,
        rxtermsDoseForm: string, 
        route: string, 
        termType: string,
        rxcui: string,
        genericRxcui: string,
        rxnormDoseForm: string,
        suppress: string, 
    };
}